Move doctrine scheme configuration from annotations to xml
