+++ config/packages/doctrine.yaml
@@ -1,16 +1,15 @@
 doctrine:
-    dbal:
-        driver: 'pdo_sqlite'
-        path: '%kernel.project_dir%/var/data/sqlite.db'
-        charset: UTF8
-    orm:
-        auto_generate_proxy_classes: true
-        naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
-        auto_mapping: true
-        mappings:
-            App:
-                is_bundle: false
-                type: annotation
-                dir: '%kernel.project_dir%/src/Core/Domain'
-                prefix: 'VDOLog\Core\Domain'
-                alias: VDOLog
+  dbal:
+    driver: 'pdo_sqlite'
+    path: '%kernel.project_dir%/var/data/sqlite.db'
+    charset: UTF8
+  orm:
+    auto_generate_proxy_classes: true
+    naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
+    mappings:
+      VDOLog:
+        is_bundle: false
+        type: xml
+        dir: "%kernel.project_dir%/src/Core/Infrastructure/Doctrine/Mapping"
+        prefix: 'VDOLog\Core\Domain'
+        alias: VDOLog
