+++ src/Core/Infrastructure/Doctrine/Mapping/Game.orm.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="https://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="https://doctrine-project.org/schemas/orm/doctrine-mapping
+                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="VDOLog\Core\Domain\Game">
+        <id name="id" type="guid">
+            <generator strategy="NONE"/>
+        </id>
+
+        <field name="name"/>
+        <field name="createdAt" type="datetime_immutable"/>
+        <field name="closedAt" type="datetime_immutable" nullable="true"/>
+
+        <one-to-many field="protocol" target-entity="VDOLog\Core\Domain\Protocol" mapped-by="game">
+            <order-by>
+                <order-by-field name="createdAt"/>
+            </order-by>
+            <cascade>
+                <cascade-remove/>
+            </cascade>
+        </one-to-many>
+    </entity>
+
+</doctrine-mapping>
