+++ src/Core/Infrastructure/Doctrine/Mapping/Protocol.orm.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="https://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="https://doctrine-project.org/schemas/orm/doctrine-mapping
+                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="VDOLog\Core\Domain\Protocol"
+            repository-class="VDOLog\Core\Infrastructure\Doctrine\DoctrineProtocolRepository">
+        <id name="id" type="guid">
+            <generator strategy="NONE"/>
+        </id>
+
+        <field name="content" type="text"/>
+        <field name="sender"/>
+        <field name="recipent"/>
+        <field name="createdAt" type="datetime_immutable"/>
+
+        <many-to-one field="game" target-entity="VDOLog\Core\Domain\Game" inversed-by="protocol" fetch="EXTRA_LAZY"/>
+        <one-to-many field="children" target-entity="VDOLog\Core\Domain\Protocol" mapped-by="parent" orphan-removal="true">
+            <cascade>
+                <cascade-all/>
+            </cascade>
+        </one-to-many>
+        <many-to-one field="parent" target-entity="VDOLog\Core\Domain\Protocol" inversed-by="children" />
+    </entity>
+
+</doctrine-mapping>
